chore(drivers/139): address share review comments
